feat：支持UI界面归档流水线和管理 #10803

+++ src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineRepositoryService.kt
@@ -1428,7 +1428,8 @@ class PipelineRepositoryService constructor(
         pipelineId: String,
         version: Int? = null,
         includeDraft: Boolean? = false,
-        queryDslContext: DSLContext? = null
+        queryDslContext: DSLContext? = null,
+        archiveFlag: Boolean? = false
     ): PipelineResourceVersion? {
         // TODO 取不到则直接从旧版本表读，待下架
         val resource = if (version == null) {
@@ -1476,22 +1477,24 @@ class PipelineRepositoryService constructor(
                 }
             }
         }
-        pipelineCallbackDao.list(
-            dslContext = queryDslContext ?: dslContext,
-            projectId = projectId,
-            pipelineId = pipelineId,
-            event = null
-        ).let { records ->
-            if (records.isNotEmpty) {
-                // 填充流水线级别回调
-                resource?.model?.events = records.associate {
-                    it.name to PipelineCallbackEvent(
-                        callbackEvent = CallBackEvent.valueOf(it.eventType),
-                        callbackUrl = it.url,
-                        secretToken = it.secretToken?.let { AESUtil.decrypt(aesKey, it) },
-                        region = CallBackNetWorkRegionType.valueOf(it.region),
-                        callbackName = it.name
-                    )
+        if (archiveFlag != true) {
+            pipelineCallbackDao.list(
+                dslContext = queryDslContext ?: dslContext,
+                projectId = projectId,
+                pipelineId = pipelineId,
+                event = null
+            ).let { records ->
+                if (records.isNotEmpty) {
+                    // 填充流水线级别回调
+                    resource?.model?.events = records.associate {
+                        it.name to PipelineCallbackEvent(
+                            callbackEvent = CallBackEvent.valueOf(it.eventType),
+                            callbackUrl = it.url,
+                            secretToken = it.secretToken?.let { AESUtil.decrypt(aesKey, it) },
+                            region = CallBackNetWorkRegionType.valueOf(it.region),
+                            callbackName = it.name
+                        )
+                    }
                 }
             }
         }
@@ -1635,7 +1638,8 @@ class PipelineRepositoryService constructor(
         userId: String,
         channelCode: ChannelCode?,
         delete: Boolean,
-        opDslContext: DSLContext? = null
+        opDslContext: DSLContext? = null,
+        archiveFlag: Boolean? = false
     ): DeletePipelineResult {
         val finalDslContext = opDslContext ?: dslContext
         val record = pipelineInfoDao.getPipelineInfo(finalDslContext, projectId, pipelineId, channelCode)
@@ -1674,42 +1678,45 @@ class PipelineRepositoryService constructor(
                         userId = userId,
                         channelCode = channelCode
                     )
-                    // 同时要对Setting中的name做设置
-                    pipelineSettingDao.updateSetting(
-                        dslContext = transactionContext,
-                        projectId = projectId,
-                        pipelineId = pipelineId,
-                        name = deleteName,
-                        desc = "DELETE BY $userId in $deleteTime"
-                    )
                     // #4201 标志关联模板为删除
                     templatePipelineDao.softDelete(
                         dslContext = transactionContext,
                         projectId = projectId,
                         pipelineId = pipelineId
                     )
+                    if (archiveFlag != true) {
+                        // 同时要对Setting中的name做设置
+                        pipelineSettingDao.updateSetting(
+                            dslContext = transactionContext,
+                            projectId = projectId,
+                            pipelineId = pipelineId,
+                            name = deleteName,
+                            desc = "DELETE BY $userId in $deleteTime"
+                        )
+                    }
                 }
-
-                pipelineModelTaskDao.deletePipelineTasks(transactionContext, projectId, pipelineId)
-                pipelineYamlInfoDao.deleteByPipelineId(transactionContext, projectId, pipelineId)
-                subPipelineTaskService.batchDelete(transactionContext, projectId, pipelineId)
-                pipelineEventDispatcher.dispatch(
-                    PipelineDeleteEvent(
-                        source = "delete_pipeline",
-                        projectId = projectId,
-                        pipelineId = pipelineId,
-                        userId = userId,
-                        clearUpModel = delete
-                    ),
-                    PipelineModelAnalysisEvent(
-                        source = "delete_pipeline",
-                        projectId = projectId,
-                        pipelineId = pipelineId,
-                        userId = userId,
-                        model = "",
-                        channelCode = record.channel
+                if (archiveFlag != true) {
+                    pipelineModelTaskDao.deletePipelineTasks(transactionContext, projectId, pipelineId)
+                    pipelineYamlInfoDao.deleteByPipelineId(transactionContext, projectId, pipelineId)
+                    subPipelineTaskService.batchDelete(transactionContext, projectId, pipelineId)
+                    pipelineEventDispatcher.dispatch(
+                        PipelineDeleteEvent(
+                            source = "delete_pipeline",
+                            projectId = projectId,
+                            pipelineId = pipelineId,
+                            userId = userId,
+                            clearUpModel = delete
+                        ),
+                        PipelineModelAnalysisEvent(
+                            source = "delete_pipeline",
+                            projectId = projectId,
+                            pipelineId = pipelineId,
+                            userId = userId,
+                            model = "",
+                            channelCode = record.channel
+                        )
                     )
-                )
+                }
             }
         } finally {
             lock.unlock()
